+++ modules/protocol/src/main/java/com/icodesoftware/Connection.java
@@ -12,8 +12,8 @@ public interface Connection extends Recoverable {
     String serverId();
     void serverId(String serverId);
 
-    long connectionId();
-    void connectionId(long connectionId);
+    int connectionId();
+    void connectionId(int connectionId);
 
     int sessionId();
     void sessionId(int sessionId);
