+++ unity-integration/Assets/GameClustering/Connection.cs
@@ -3,7 +3,7 @@
     public class Connection
     {
         public string Type { set; get; }
-        public long ConnectionId { set; get; }
+        public int ConnectionId { set; get; }
         public int SessionId { set; get; }
         public string Host { set; get; }
         public int Port { set; get; }
